backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
starlette==0.27.0

# Deep Learning & NLP (Improved Models)
torch==2.1.1
transformers==4.35.2
tensorflow==2.14.0
scikit-learn==1.3.2

# Data Processing
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4

# API Clients & HTTP
requests==2.31.0
httpx==0.25.1
aiohttp==3.9.1

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging & Monitoring
python-json-logger==2.0.7

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Async & Concurrency
aiofiles==23.2.1

# Testing (Development)
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0

# Additional for API integrations
beautifulsoup4==4.12.2
lxml==4.9.3
